Fix over 10 HTML validation errors, but breaks current functionality.
This commit fixes over 10 HTML validation errors. These errors were
caused by the following bugs:

* `$` char in HTML ID value. Only alphanumeric, `_`, and `-` are legal.
 * For the time being, I changed this to `_`, but this breaks the JS.
* Empty `<table>` and children. No stub elements are allowed.
* `<input>` elements were not self-closed. They must be for our DOCTYPE.

As of this commit, the `themeChooser()` JavaScript functionality is not
working because it relied on invalid HTML and, having fixed the HTML,
its dependency is broken. The HTML is still not entirely valid, but is
better.

NEXT STEPS:

1. The HTML needs to be validated and semantic.
2. Non-JavaScript functionality should be added to provide graceful
    degradation.
3. Any necessary JavaScript should be added last.

I'm not intimately familiar with what the exact functionality is
supposed to be, so I'm at a bit of a block trying to figure out exactly
how the non-JavaScript version of the HTML should really look.

git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@15877 e846bd91-c042-42a7-a218-4b65e2c1904d