Fixes RT #24858 for internationalized wikis.
This applies the previous fix for English-language wikis across all
Wikia wikis regardless of natural language. The pro to this is, of
course, that it works across all languages, but the con is that the fix
for IE7 is now coupled with the Monaco skin. However, since other CSS
hacks in extensions already couple themselves with the skin, this is no
worse for the situation as it currently stands.

git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@16219 e846bd91-c042-42a7-a218-4b65e2c1904d