Fixes RT #33649: Return Activity Feed elements to normal flow.
This commit reworks the way sprite images are handled. Previously, a
number of Activity Feed elements were required to be relatively
positioned. This causes problems such as described in RT #33649.

This introduces a `getSprite()` function to the `FeedRenderer` class
that is responsible for outputting an appropriate sprite for a feed and
removes no-longer necessary CSS. It also begins moving sprite CSS to the
skin's root style sheet so it can be shared across extensions.

git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@16379 e846bd91-c042-42a7-a218-4b65e2c1904d