(BugID: 52365) Add URL fragment to the link cache key
Rendered links will be cached to use the same URL as the first link
to a particular page meaning that when more than one link on a page
links to the same page but different sections (i.e.
[[Test page#Section_1]] and [[Test_page#Section_2]]) then both links
will be rendered with the same section (the section used in the first
link to that page).

Adding the URL fragment to the cache key solves this issue.