(BugID: 69180) Fix inability to remove global rights
Special:UserRights on submission checks against rights from the local
database when checking which rights to remove. This uses the loadGroups
method in the User class which isn't touched by ShareUserrights. As the
global group isn't returned by this method, no global groups are
determined as removable. To fix this, hooked into the loadGroups method
to include global groups.