Fixed problem with adding images on non EN wikis.
Problem is that the code is matching the [[File...]] tag in the content language when its looking for the right tag to replace.  If they've used english rather than their content language when adding these tags, then the old code wouldn't find it.  Now match for these tags in both content language an English.

BugID: 4935