(BugID: 95202) Add interwiki prefix to the link cache key
Rendered links are cached to use the same URL as the first link to
a particular page. As the cache key didn't take into account the
interwiki prefix used, if two or more interwiki links pointing to
different sites had the same page name and link text, they would
all have the URL of the first interwiki link.