(BugID: 95531) Fix PHP error in WikiaRSS extension
A PHP error was caused when the charset option was not used with
both the full and short options:
mb_convert_encoding(): Illegal character encoding specified

This was caused by an incorrect empty check in the getFullItemList
method and no check in the getShortItemList method.

Also replacing the use of wgOutputEncoding which was removed in
MW 1.19 and now everything uses UTF-8 regardless.