Truly associates category data to the DOM.
Fixes not correctly de-duping categories with HTML entities. No longer
have to re-parse articles when loading CategorySelect. When saving
categories on article pages, the view is updated with a more accurate
DOM structure since it uses the same mechanisms as the original article
view request. Category types are now added to category data so it can be
used for styling purposes. Among other things, fixes BugId:97734.