(BugID: 97137) Fix SiteWideMessage notification in monobook
Moving SiteWideMessages monobook notification into the SiteNotice.
This makes it more obvious that the user is seeing an SWM, rather
than the broken You have new messages link. This also displays the
message on every page so it behaves similarly to the notification in
the Wikia skin.

This also simplifies the monobook notification code by replacing
six hooks with one.