(CE-75) Recent changes filter not respecting "All" preference
If you set the recent changes namespace filter from selected namespaces to
all, then visit another wiki with different namespaces, the differing
namespaces will be unselected. This is because the namespaces selected in
the dropdown are saved for the user rather than clearing the option to
get back all being selected properly.