(CE-41) Blocked users should be able to post on their own Wall
Fixes the previous fix. getBaseText was returning the whole title rather
than the first part of the title (the username) because the Thread
namespace does not have subpages enabled. Even it does, replies to
other comments will break anyway since it returns everything before the
last /. This forces it to always get the base page name which will be the
username.