Small fixes and cleanup for EditAccount::closeAccount
No need to pass a title object into this method, since the title will
always be the special page name. The second status message wasn't
getting set on failing to remove the user's avatar. Reorder the
parameters of the method as the user account being closed should
logically be the first parameter of this method rather than the
change reason.