(BAC-1198) Fix use of explode in getData
The explode method which expects a string was being passed a Title
object. This worked for loading comments on Articles, Blogs, Walls,
and Forums because of Title's toString method (every other use of
explode correctly passes a string to it). This failed on some
WikiActivity pages however and instead of returning a string in
$parts['title'] it returned the title object.

We should be correctly passing a string of the title without a
namespace, rather than relying on toString, and then constructing
the subject page using the $parts['title'] and the subject namespace.