VID-1515 Launch agegate when player is loaded
This is code from Ooyala. To fix the issue with iPhone's sometimes not
closing the video player which leaves the agegate stranded in the
background. Instead, we launch the agegate as soon as the player is
ready, rather than waiting for the user to click the thumbnail. Note,
preloading the agegate like this only occurs on mobile.