(CE-769) Don't try to convert the birthdate into a timestamp
The user_birthdate field is a date field, but in these two methods
we were using timestampOrNull which will return false when not given
a valid timestamp. So, when a date was passed to it instead of a
timestamp, false would go here, which would result in a user_burthdate
of 0000-00-00. So, by removing the timestampOrNull call, we match the
way we handle this in ExternalUser_Wikia.php which is what will be used
most of the time.