CE-377 Decouple UserProfilePageV3 from Phalanx
UserProfilePageV3 had an implicit dependency on Phalanx, which prevented
it from working as expected on Wikia One (which has UserProfilePageV3
enabled and Phalanx disabled). Extract the determination of the content
type from UserProfilePageV3 into Phalanx, which now uses the class of
the given subject to determine the type if none is given.