CONN-512 move doDeleteUpdates after the hook
`doDeleteUpdates` has [side
effects](https://github.com/Wikia/app/blob/CONN-512/includes/WikiPage.php#L2129-2133)
that effectively reset the page. This disables hook logic that
needs to manipulate the page being deleted or it's associations. In this
case, we need update the comments index row that is associated with this
page and mark it as deleted.

This fix is brittle. IMHO there should be FK constraints from
`comments_index.first_rev_idi` -> `revision.rev_id` that cascade on
delete.