(CE-815) Fix Scribunto fatal error when parser is cloned
Fixes a fatal error in Scribunto when the parser is cloned, such as
when DPL is used on the same page as a Lua module. This allows us to
give the new Parser instance a new ScribuntoEngine, as otherwise the
engine for the other parser will be unexpectedly destroyed. Scribunto
already makes use of this hook to handle this, so we just need to
backport the ParserCloned hook call.