PLAT-170 moving WikiaLogger to lib/wikia
Our logging functionality is core to the functioning of the platform. As
such it shouldn't be loaded as an extension but a library. This commit
moves the WikiaLogger to lib/wikia.

As part of this move, I'm also moving Monolog and Psr to composer so we
don't have to handle the autoloading.

/cc @nmonterroso