Move ES log toggling from global to class property.
Moved ES log toggling from global variable ($wgDevESLog) to class
property on the formatter. As part of the move, added DevModeFormatterInterface to
facilitate the toggling. Now, in order to use ES/logstash in dev, you
need to toggle the formatter and enable dev mode.

  \Wikia\Logger\WikiaLogger::instance()->getSyslogHandler()
    ->setModeLogstashFormat()->disableDevMode();

I also moved the setup of the error handler and the toggling of dev mode
to Wikia\Logger\Hooks. This isn't quite right because it means that some
messages will be logged locally in production until this hook is loaded.
It's a placeholder for now.