PLATFORM-75 Make TaskRunner aware of the task ID
For longer-running tasks, a timeout somewhere between application/dev
boxes and the celery workers can cause the task status to become
"failed" when it could simply be running longer. Passing the task ID
along to the runner will enable it to communicate back to Flower that
the task actually did complete successfully.