PLAT-170 Log sample of SQL queries w/ WikiaLogger.
Inserted random (at a rate of 1%) sampling of SQL queries in
DatabaseBase around doQuery. The instrumentation adds the query, the
execution time, the function, and the server role to the log payload.

With this in place, we can query using Kibana for the URL, the query
string, the server, the db name the calling method:function, and the
elapsed time.