PLATFORM-275 Invoke UserRenameTask on default wikiCity only
Attempting to invoke the UserRenameTask on each wikiCity individually would
fail when the target wikiCity did not have the UserRenameTool extension
enabled, which should be all of them except "community". Instead, run
the task on the "community" main wiki and shell out to the pre-existing
maintenance script that bypasses the enabled check by force-requiring
the dependencies.

Also, better logging.

Also,
PLATFORM-75