PLATFORM-75 Deprecate UploadFromUrlJob
This job requires `$wgAllowAsyncCopyUploads` to be `true`, but the
definition in DefaultSettings.php warns "This feature is experimental
and broken as of r81612." That warning was written in 2012, so it is
very likely that this has never been enabled (and would have caused a
number of things to break if it was attempted).

Deprecate the `$wgAllowAsyncCopyUploads` variable, remove the
`UploadFromUrlJob` class, and remove most (all?) references to async
uploads.