PLAT-75 WIP Refactoring HAWelcom job into a task
This job migration is a work in progress. I'm breaking the hooks and the
class declaration out into separate files and adding a .setup.php file
for the extension. This should move the extension closer to the current
extension format.

I'm also adding tests in attempt to smooth the refactoring efforts.

/cc @nmonterroso @michalroszka