PLAT-345 ensure sender object is passed to doEdit
It appears that the sender object has been missing from the parameter
list to Article::doEdit which is responsible for the work of creating
the welcome message. As a result, the $wgUser is being used as the
default author in a batch context which appears to resolve to localhost
or 127.0.0.1 in this context.