PLAT-345 pass $user param further down the stack
ArticleComment::doSaveAsArticle was not passing the $user object to
EditPage::internalAttemptSave which uses the $wgUser global by default.
As a result, the HAWelcome wall messages were being posted as the
"defualt" $wgUser when it's not set which is 127.0.0.1.

To fix this, I added a $user parameter to internalAttemptSave and pass
it down the call stack.

Additionally, instead of using the senderObject, we use the default welcome
user when posting to the talk page. The senderObject is still used to
create the message payload.