POC helper for wrapping global state
This commit adds a GlobalStateWrapper class that wraps global variables
with values specified in the constructor that will be used to create a
context within which a supplied anonymous function can be called. The
state before the function call will be restored after the call
completes. Exceptions are also captured and can be extracted from the
object if the caller so desires.