PLAT-345 wrap $wgUser object when posting to wall
The call stack used by buildNewMessageAndPost does not pass the $user
object all the way down the stack but instead falls back to $wgUser. We
have two choices-- we can alter core to pass $user down or we can wrap
the global. The former is preferred but risky. The latter is a compromise with the
GlobalStateWrapper. Going with the latter.

/cc @owend