Use cached data if available
To increase performance; check the cache for existing data
if there is data use it. This will prevent calls to the Monetization
Service if it isn't required.

NOTE: when ecommerce modules are added we will have to change the logic
because the cache key is site/geo specific and the products to be
displayed are specific to the page. The cache key will have to be
site/geo/page id.