(CE-1000) Assign user page creations to the default welcomer
Prior to the rewrite, the user page creations were assigned to the
default welcome user (User:Wikia) due to the overwrite of $wgUser
global rather than passing it to the doEdit method. This restores the
expected functionality by passing the default welcome user to doEdit
rather than the sender object.