VID-1946 Removed unused lazy load attributes
We have a number of different attributes which we only set if we're using
(controller->noscript, controller->rdf to name a couple). If we
decide to declare all of them ahead of time, then we should do that,
otherwise it seems like it would be better to be consistent and only
add things if they're actually used.