Replace thumbnail generation with vignette
Replace the thumbnail generation with vignette in the File abstraction.
This is more complicated than expected because the timestamp used to
store the archive file on disk is not the timestamp stored in the
oi_timestamp column in the oldimage table. The timestamp used on disk is
stored in the file name.
 - add getArchiveTimestamp to the file abstraction so we can extract
   the timestamp from the file name.
 - add isOld to the FileInterface to distinguish between archive
   images and current images
 - adjust the UrlGenerator to be aware of old vs not image so that
   the timestamp can be generated properly
 - add tests for the new behavior in UrlGenerator