(CE-937) Use staff welcome message for non-admin Helpers
Make helpers who are not local admins use the staff welcome message
when welcoming users. Some helpers are listed as the staff member for a
language which is used as a fallback to determine the welcomer when
there was no recently active admin. In those cases, they should use
the staff welcome message so they are not confused for a local admin.