(CE-1143) Load JSSnippets in Monobook
In PR #5237, loading of JSSnippets was moved from a hook to the
AssetsManager config, but wasn't added to load in monobook_js as well.
This caused a number of JS issues in Monobook for extensions that made
use of JSSnippets, so we just need to add JSSnippets to monobook_js.

This fixes the following tickets:
* CE-1143: Tabber is broken in the Monobook skin
* CE-1144: RSS feeds still broken in Monobook
* CE-1147: Polls in Monobook are not working