(CE-1051) Quick fix for caching in SharedHelp
SharedHelp tries to get the parser cache key from the text it gets
from the help wikia, and use that to save HTTP requests. The string
in the text it is parsing looks like:
<!-- Saved in parser cache with key wikia:pcache:idhash:16765-1!*!0!!en!*!* -->

Shared help was looking for the string 'end' to find the end of the
cache key. Because this string isn't in that comment, the resulting
parser cache key it would store would look something like
'wikia:pcache:idhash:16', which of course doesn't work. I'm not sure
this ever worked and probably should have been 'and' in the previous
version of the string before we removed the timestamp. This fixes that
parsing so the working parser cache key is stored.