Remove $wgEnableUserLogin variable and unused code.
This change is one of many cleanups along the way of the Helios project.

`$wgEnableUserLogin` is set to on the config level `true` with no
exceptions. As Wikia's UserLogin extension is now a critical part of
user-related flow in the application, noone is going to remove it.

As a matter of fact, the extension is enabled even for Uncyclopedia and
Wikia One.

Parts of the extension will be refactored as the authentication project
(Helios) progresses.

There is no need to check for `$wgEnableUserLogin` and no need to keep
code that will never be executed or wrap code that will always be
executed.