VE-1733: Fix errors related to toolbar or toolbarButtons being not defined by correcting order of calling certain methods
In this PR https://github.com/Wikia/app/pull/5616 call to method setupToolbarButtons and attachToolbarButtons got deleyed by introducing afterHideSpinner which caused few different problems. One problem was related to sanity check - which in the outcome needs to change state of the save button, second problem was related to focus widget and it dependsnecy on toolbar in onDocumentSetup method.

This is release blocker that's why commiting directly to "dev".

https://wikia-inc.atlassian.net/browse/VE-1733