(CE-1192) Fix creating new pages with forms on some SMW wikias
SemanticForms was doing a hacky deep clone of the parser to get around a
bug I can't reproduce, by serializing and unserializing it and then using
the clone to parse the form. This caused an exception as parts of the parser
contain closures which cannot be serialized. So backporting the following
change from a newer version of the extension to fix it:
https://github.com/wikimedia/mediawiki-extensions-SemanticForms/commit/ca8d126424c1

Fixes fatal error:

    Unexpected non-MediaWiki exception encountered, of type "Exception"
    exception 'Exception' with message 'Serialization of 'Closure' is not allowed'
    in extensions/SemanticForms/includes/SF_FormPrinter.php:428