Initial implementation.
We determine whether Helios is enabled or not from the global variable $wgEnableHeliosExt (enabled if true, otherwise disabled).

We determine whether the shadow mode is enabled for Helios logins from the global variable $wgHeliosLoginShadowMode (shadow mode if true, otherwise the normal mode).

A dicrepancy between MediaWiki's and Helios' authentication results indicates and error and we must know about it.