(CE-1223) Fix UserSignup unit tests
Fix UserSignup unit tests after changes. Most of these tests were written
when we had the temp user setup, and need cleanup. This removes those that
are not relevant such as the first two which were exactly the same, and
changes the expected result for when an invalid user is provided to match
the new behaviour.