MAIN-3427: Use keypress instead of keyup event to handle "enter key" in title input box
This solved the problem which was occuring within IME input (for instance Hiragana). In IME while typing "enter key" sometimes have to be used to accept composition - and it triggers keydown, keyup, but no keypress. As previously handler was attached to keyup instead of accepting composition dialog was being closed.

https://wikia-inc.atlassian.net/browse/MAIN-3427