Helios/Client - use GlobalStateWrapper for HTTP requests to Helios
MediaWiki's MWHttpRequest heavily relies on wfMessage() and therefore
requirers $wgLang to be set. $wgLang is set after $wgUser and sometimes
(for per-request authentication with access token) we need the variables
to be set in the opposite order. GlobalStareWrapper makes it very simple
and locally replaces $wgLang with $wgContLang which is set before $wgUser.