PLATFORM-1232 force a FromSession user in Piggyback sessions
When an access token is provided, the user object for the request
is created based on that. It makes Piggyback useless as Piggyback
relies on session data accessed later during the execution.

We can distinguish a Piggyback session from a regular one by
a simple check and override the user object created from an
access token with one created from session data.