SOC-1075 Make sure cache is cleared for the old user ID
Previously, user data was loaded from cache to delete user data from cache.  This ridiculous loop meant that the very thing we were trying to clear was overwriting the new data we were trying to write.

Also quite a bit of code clean up was done while trying to understand what was going on in the rename code.