(CE-2623) Allow editing of MW JS pages when ContentReview is enabled
JS pages are now blacklisted from Verbatim, and the JS editing protections
are skipped for MediaWiki JS pages if ContentReview is enabled, so we
need to also whitelist them so those with editinterface but not
editinterfacetrusted can edit them. The extra check if ContentReview is
enabled is for paranoia as it should already be checked in the ProtectSiteJS
extension.