Verify helios token when load user from the session
Verify the helios authentication token when loading a user from the session. If
we somehow get to User::loadFromSession then we would also verify that
the user has a valid authentication token.

This *should* ensure that if a helios authentication token is revoked
that a user will no longer be authenticated regardless of the MediaWiki
authentication mechanism(s).