(SEC-14) Don't process message further when replacing rawParams
In https://github.com/Wikia/app/pull/8457 we added the ability to
supply rawParams to mw.message so that they would be passed after
the message is escaped. But if the jqueryMsg library is loaded, the
current implementation means this becomes unescaped because rather
than using the parser in mediawiki.js, it uses the parser in jqueryMsg
again. The jqueryMsg parser actually unescapes text because it
expects mediawiki.js to escape the text after the parser is run, which
we're running again after escaping.