VE-1975: Added support for mobile (ArticleAsJson) and usage of XML parsing
Special handling for mobile is needed because article for mobile is delivered within ArticleAsJson which makes a lot of transformation during the parsing process (instead of for instance postprocessing).
Parsing HTML with XML parser seems to be giving better results than using just regexes and also it gives as ability to handle nested tabbers in a future in a better way.