VE-2004: Silencing loadHTML warning
I carefully researched the topic and it seems that there is no way to instruct DOMDocument to accept "lower" quality HTML. Specifically in HTML <div> or <figure> (block elements) are not allowed inside <p> (which accepts only inline elements) - hence it's complaining about.

Based on comments from stackoverflow it seems that such practice (silencing it) is commong practice in this situation.