VE-2069: Adjust getImages method to work well with recent changes to ImageNode
I'm not very happy with how nested this method is but it seems that it is the convention taken within PortableInfobox project.
I would much better prefer if there was more object oriented way to delegate for instance task of retrieving images list directly to NodeImage but that would be much greater refactoring.