ForeignAPIRepo: fetch thumbs directly from Wikimedia
ForeignAPIRepo was trying to fetch and store thumbs on our DFS and use file-based locks on `/images`. This obviously did not work in our environment and caused a fallback to fetching thumbs directly from Wikimedia.

Solution: change the $wgForeignFileRepos config that handles $wgUseInstantCommons setting to skip the thumbs caching step.