SEO-181 Fix removal of hreflang link to self
The code checks the language code and removes the link from the
structure based on that language code, so there's no link to an English
article on an English article (no point linking to self).

The problem is the code checked the user language instead of the content
language and this commit fixes it by fetching the language from wgTitle.