(CE-3331) Fix block info in the userinfo API module
The blockedby value was returning false for local blocks as User::blockedBy()
returns the user name for local blocks, rather than the ID. So, use the Block
object instead.

Also set a blockid value in the response, which will be a better check for
whether the current user is blocked.