SUS-327 Fix for polls on forum and wall threads not rendering correctly
The polls on forum and wall threads don't load the needed CSS or JS
files if they are the only content for the message, example:

    <poll>
    What's your favorite color?
    red
    green
    blue
    </poll>

If you surround the poll with some wiki text it's gonna work though:

    Here's some poll:

    <poll>
    What's your favorite color?
    red
    green
    blue
    </poll>

    Let the others know!

I figured, for some reason passing just the code generated by the
AjaxPoll to `wfFixMalformedHTML` strips down the HTML comments and the
`<script>` tag with the `JSSnippetsStack.push` line (that line is
responsible for loading the Polls CSS and JS).

This change makes the input passed to `wfFixMalformedHTML` wrapped in a
`<div>`, that's later stripped from the result of the function.

It fixes the issue.